:bug: Fix the bug when the ja_shop_name is empty.
We don't know why we've got this, but now we check for the current ja_shop_name field in the database and we have a manual process to deploy a new server without problems.

Now we can have 3 different behavior on the product's name:
- Add "- Vendor name" if everything is okay and the incoming field from Splash isn't empty.
- Add "- Vendor name" but from the current object. If the incoming field is empty but the current field is not, we use it!
- Add "- ERR" if the incoming and current fields are empty.

In every case, we can retrieve all new products in the splash_temp category and we can see what's wrong from here.